remove env vars, simplify
